+++ TerminalDataBase/TerminalDataBase/Identifiable.h
@@ -4,5 +4,5 @@ class Identifiable
 {
 public:
     virtual ~Identifiable() {};
-    virtual const __int64& getId() const = 0;
+    virtual const long long& getId() const = 0;
 };
